+++ assets/scripts/desktop/install-chrome.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -Eeuo pipefail
+
 cd "$HOME/.cache/dotfiles/"
 
 wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb